compute recommended threshold for training images
I found experimentally that a nose+mouth usually has a >98% correlation with a solid white background. However, the same image usually has a >99.8% correlation with an actual nose+mouth. I did not feel that it would make sense to hard code these exact values, because it probably varies with the image's dimensions and exposure.

To deal with this, I compute the dot product of each training image with a solid white background. I then average this with 1.0 (finding the "middle ground" between solid-white correlation and 100% correlation) and use it as the recommended threshold. This seems to work experimentally for images of myself, but I have yet to test it with large training sets.