bail early for low correlations in template.correlation()
By taking a threshold argument, template.correlation() can return early if there is no way it could exceed the threshold. To tell this, we compute the best possible correlation coefficient each time we go through the y-coordinate loop. If this correlation coefficient is less than the threshold, we bail.

To figure out the best possible remaining dot product, I used some simple linear algebra. Simply put, we need to maximize (v dot r), and we know the magnitude of r (since we know the total magnitude of the image sub-region in advance, and we can subtract off bits of that as we go). With that in mind, the optimal value for r turns out to be v*||r||/||v||.