moved optimizations into new subregionInfo type
This makes it a lot easier to read the optimized code, especially because it removes the need for tons of arguments to the template.correlate() function. In general, introducing a "subregionInfo" type seems pretty intuitive, since it gives us a way to refer to the current subregion of the image.

It should be noted that this change has a minor performance penalty, but it is worth it for more readability. Plus, it may allow for more optimizations in the future.