let the user customize the threshold strictness
The way I see it, there are two possible ways to get successful face recognition, and both of them require different strictness values. One way is to have accurate pictures of many mouths and many eyes; then, all you must do is show a mustache wherever there is either a mouth or eyes. This increases the probability of finding a face by a large factor, but it makes weeding out false positives even more difficult. For this, you would want a high strictness.

The other technique is to use several general eyes, mouths, and other features, and set a low threshold on these features. Then, when it comes time to look for faces, you can look for places where there is both a mouth and a pair of eyes. This way, you need less training samples (and the algorithm can run faster). However, this requires that you successfully match pretty much every mouth and every pair of eyes, which might lead to many false positives. For this, you would want a low strictness.