refactor mustacher API
The Template type is now the standard way of storing a Database entry; there is no longer a private type for it.

Now, there aren't separate Correlation and DatabaseMatch types. Instead, the Match type embodies both of these types. This makes it possible to get all matches from a database, or even from an ElasticSearch, and then filter at your own discretion.